- separate opacity in 2
- image opacity only for background image
- background color opacity only effects color
- use opacity image also in overview

+++ EmojiArt/Detail/ColorRgbModal.swift
@@ -10,5 +10,5 @@ struct ColorRgbModel: Codable {
     var red: Float
     var green: Float
     var blue: Float
-    var alpha: Float
+    var alpha: Double
 }
